market/requirements.txt
flask==3.1.0
gunicorn==23.0.0
apify-client>=1.8.0
tavily-python>=0.5.0
openai>=1.50.0
python-jose[cryptography]==3.3.0
flask-cors==4.0.0
